// src/app/models/newSocioEcoSituation.model.ts
export interface NewSocioEcoSituation {
    dependentMembers: number;
    totalIncome: number;
    housing: string;
    tenancy: string;
    numberOfRooms: number;
    totalExpenses: number;
    waterConsumption: number;
    electricityConsumption: number;
    livingPlaceConsumption: number;
    feedingConsumption: number;
    educationConsumption: number;
    telephoneConsumption: number;
    transportationConsumption: number;
    healthConsumption: number;
    otherConsumption: number;
    waterConsumptionDetail: string;
    electricityConsumptionDetail: string;
    livingPlaceConsumptionDetail: string;
    feedingConsumptionDetail: string;
    educationConsumptionDetail: string;
    telephoneConsumptionDetail: string;
    transportationConsumptionDetail: string;
    healthConsumptionDetail: string;
    otherConsumptionDetail: string;
}